[fix unitest]
