+++ configs/classification/timesnet_SelfRegulationSCP2.yaml
@@ -7,12 +7,6 @@ epoch: 10 # max_epochs
 training: True # 
 eval_metrics: ["acc", ] # 
 
-lr_scheduler: # 
-  name: Type1Decay
-  lr_cfg:
-    learning_rate: 0.001
-    last_epoch: 0
-
 dataset: 
   name: SelfRegulationSCP2
